#2 Stub out branches for -run-etl-process & -write-etl-results-to-file
This patch implements the top-level arg parsing of the ETL processes we want to support.

We now have stubs for the following:
* ./arxiv_as_a_newspaper.py -run-etl-process
* ./arxiv_as_a_newspaper.py -write-etl-results-to-file <destiation.json>

These stubs raise NotImplementedError instances.

Actual support for these functionalities is not included in this patch.

This patch also includes nice error handling and a lot of descriptive error messages.