#2 Change access privileges on "Recent" Pages scraping utilities
A lot of the functionality in the "Recent" Pages scraping utilities isn't useful outside of the public interface to our ETL utilities, so we chose to make those not accessible outside of the modules in which they are defined.

The functionality changed include:
* recent_page_url_info_tuple_to_json
* extract_info_from_recent_page_url
* append_abstract_to_info_extracted_from_recent_page_url
* abstract_text_from_arxiv_paper_url
* extract_info_without_abstract_from_recent_page_url
* extract_info_tuple_iterator_from_recent_pages_definition_list
* extract_info_tuple_iterator_from_definition_term_description_doubles
* extract_info_tuple_from_definition_term_description_double
* author_to_author_link_double_to_author_to_author_link_dictionary

Due to the renames / access privilege changes, we had to accordingly change callers as well, e.g. extract_info_from_recent_page_url_as_json

This patch also includes a documentation extension on extract_info_from_recent_page_url_as_json.

We also added missing imports.