#2 Rename utilities/etl_utilities.py -> utilities/extract_transform_utilities.py
It was decided in https://github.com/paul-tqh-nguyen/arxiv_as_a_newspaper/issues/2 that we should separate out the extraction and transformation (i.e. the https://arxiv.org/ scraping) functionality from the loading functionality (i.e. the functionality that writes to our MongoBD).

This patch takes the first step in doing that by renaming the modules.