#2 extract_transform_utilities.py to return dicts instead of JSON
As we've completelyed our load utilities, we've realized that there's no need for us to write functionality that converts our extracted and transformed data into JSON. We need to merely internally represent them as Python dictionaries and pymongo will handle the rest. Thus, we've updated the utilities in extract_transform_utilities.py to return Python dictionaries instead of JSON strings.