#2 Spread URL requests across mirror sites.
This patch makes it so that when we're scraping arXiv, we also hit their mirror sites so that we don't get blocked after a while.

This also makes it so that we wait some number of seconds before hitting a URL (so that we can slow down how often we're hitting URLs).