#8 Update ETL to also include overall research field when uploading new docs to the DB
When the ETL process loads docs to the the DB, it does it in batches where the contents of each batch are currently extensionally defined to be all the research papers relevant ot a general research field, e.g. Mathematics, (as opposed to a specific field, e.g. Combinatorics). We now make it so that the documents loaded to the DB also include this information.