#12 Hook up front end to top-level CLI.
This patch makes it so that the following now install the necessary libraries to get our front end started and starts our front end server:

                            ./arxiv_as_a_newspaper.py -start-front-end-server

It also has some nice pretty printing.

It also gracefully handles keyboard interrupts, which is (as the printing shows when we invoke our front end as described above), the preferred way of exiting the front end.