#20 Initial Commit For Header Row
This is the first pass at making the header row (part of the front end interface) functional.

We have a set of reasearch fields that are displayed in the header row.

They switch out every 3 seconds.

They fade in and out.

There are 6 shown at a time (or less if we're at the end of the list of research fields).