#28 Make the Center Frame Reasonably Aesthetic
This finished up #28.

This makes it so that the articles shown in the center frame are somewhat pretty.

We added a bunch of fonts.

We added lines between articles.

With styling changes, there isn't usually much to say in words.

There's some relevant cheesey quote regarding words and pictures that's relevant here.

We added a numch more CSS classes and IDs.

We refactored the components for the center frame some.

There's much more refactoring to be done there.

This is a progress patch in the grand scheme of things.

there'll be much more refactoring to come in the future.