#23 Initial Commit Of Side Navigation Bar
We want to be able to have all of the available research fields visible at once.

We plan on doing that by having a side navigation bar showing all the research fields at once.

This patch commits the necessary react components to do so.

There's still a bunch of missing CSS, which is what we'll have to do next to finish up #23.