#38 Update the favicon to match the aesthetics of the front end
This patch updates the favicon.ico to be aesthetically more in line with the aesthetics of the page.

The old favicon was maroon, which is in line with the aesthetics of the original arXiv website, but not with out front end.

Our front end is a mostly monochrome page, like that of a newspaper.

This patch makes the icon an AX, representing ArXiv, in black and white.