o#32 Make loading screen transition more smoothly
This includes several changes to the loading screen.

We no longer display the "Loading" text as it was deemed distracting.

We now smoothly have the loading screen appear for a few seconds before disappearing without any hard changes or hits.

The color scheme now matches that of the actual website.