#47 Close Side Navigation Bar On Outside Click
This makes it so that if we click outside of the side bar when it's open, we close the side bar.

It was not intuitive how to close the side bar, and the obvious link to close it was all teh way on the otherside of the screen.

This also makes it so that clicking a side bar link closes the side bar.