#52 Initial commit of "-deploy" CLI option support
This patch includes the initialy support for a "-deploy" option in our top-level application.

This'll make it nicer for us to have a single button to push to publish all of our changes.

It's pretty simple.

I just calls out to "npm run deploy" in the relevant front end directory.