#58 Fix Side Bar Missing Items
The CSS made it such that we weren't displaying all the elements of our side bar since the side bar actually ran off the bottom of the page.

This was because we set the height of the side bar to 100%, but the top of the side bar started at 48px.

Thus, 48px of the side bar was off the bottom of the page.

This patch updates the CSS so that the height of the side bar is 100%-48px.

We could tell that we were missing items bc the "Next Research Field" button worked when we clicked the bottom most visible research field in the side bar