#54 Appropriately hide "Next Research Field" and "Previous Research Field" when not applicable.
This patch makes it so that we don't display the "Next Research Field" link / button when there's no next research field to go to and the vice-versa regarding "Previous Research Field".

We now pass around the entire research field list to the relevant subcomponents in order to be able to detect when we're at the end or beginning of the research field list.