#59 Add permanent scroll bar
Much of the aesthetics of our front end functinoality depends on relativate placement, e.g. center justification and such, so the addition or removal of a scroll bar makes it such that the relative placements of page elements change.

Since the amount of text on a page changes when we shift research fields, such events can cause the aforementioned phenomenon.

This can be prevented if there's a permanent scroll bar.

This patch makes it so.